requirements.txt
Django==2.0.7
django-bootstrap3==10.0.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
pytz==2018.5
